# contracts/openapi/candidate-service.yaml
openapi: 3.1.0
info:
  title: CDC ATS — candidate-service
  version: 0.0.1
  description: >-
    Candidates, applications (incl. hire/reject/advance), offers + offer approval
    chain, three AI agents (sourcing, offer, candidate-experience), CSV/Excel
    import, and the interview-question generator (default port 4005). All routes
    mount under /internal/*. RLS-scoped request client. Publishes
    tenant.{t}.application.hired / .rejected and tenant.{t}.offer.approved /
    .accepted (lib/decision-events.ts). Subscribes to tenant.*.resume.parsed,
    assessment.completed, interview.round.started.

    Gateway mapping: /api/candidates -> /internal/candidates; /api/applications ->
    /internal/applications; /api/offers -> /internal/offers; /api/sourcing (AI) ->
    /internal/sourcing; /api/offer (AI) -> /internal/offer;
    /api/candidate-experience (AI) -> /internal/candidate-experience.
servers:
  - url: http://localhost:4005
    description: Internal service port
...
paths:
...
  /internal/candidates/{id}/applications:
    get:
      tags: [candidates]
      summary: A candidate's applications
      parameters:
        - $ref: "./_shared.yaml#/components/parameters/XUserId"
        - $ref: "./_shared.yaml#/components/parameters/XTenantId"
        - $ref: "./_shared.yaml#/components/parameters/XUserRole"
        - $ref: "./_shared.yaml#/components/parameters/XInternalService"
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Applications., content: { application/json: { schema: { $ref: "./_shared.yaml#/components/schemas/SuccessEnvelope" } } } }
...
  /internal/applications:
    get:
      tags: [applications]
      summary: List applications (filter by requisition / stage)
      parameters:
        - $ref: "./_shared.yaml#/components/parameters/XUserId"
        - $ref: "./_shared.yaml#/components/parameters/XTenantId"
        - $ref: "./_shared.yaml#/components/parameters/XUserRole"
        - $ref: "./_shared.yaml#/components/parameters/XInternalService"
        - { name: requisitionId, in: query, required: false, schema: { type: string } }
        - { name: stage, in: query, required: false, schema: { type: string } }
      responses:
        "200": { description: Applications (includes applicationsByStage source)., content: { application/json: { schema: { $ref: "./_shared.yaml#/components/schemas/SuccessEnvelope" } } } }
    post:
      tags: [applications]
      summary: Create an application (ADMIN/RECRUITER/HIRING_MANAGER)
      parameters:
        - $ref: "./_shared.yaml#/components/parameters/XUserId"
        - $ref: "./_shared.yaml#/components/parameters/XTenantId"
        - $ref: "./_shared.yaml#/components/parameters/XUserRole"
        - $ref: "./_shared.yaml#/components/parameters/XInternalService"
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        "201": { description: Created., content: { application/json: { schema: { $ref: "./_shared.yaml#/components/schemas/SuccessEnvelope" } } } }
        "400": { $ref: "./_shared.yaml#/components/responses/ValidationError" }
        "403": { $ref: "./_shared.yaml#/components/responses/Forbidden" }
  /internal/applications/time-to-hire:
    get:
      tags: [applications]
      summary: Time-to-hire metric series
      parameters:
        - $ref: "./_shared.yaml#/components/parameters/XUserId"
        - $ref: "./_shared.yaml#/components/parameters/XTenantId"
        - $ref: "./_shared.yaml#/components/parameters/XUserRole"
        - $ref: "./_shared.yaml#/components/parameters/XInternalService"
      responses:
        "200": { description: Time-to-hire data., content: { application/json: { schema: { $ref: "./_shared.yaml#/components/schemas/SuccessEnvelope" } } } }
  /internal/applications/attachments:
    post:
      tags: [applications]
      summary: Record file-attachment metadata (ADMIN/RECRUITER/HIRING_MANAGER)
      parameters:
        - $ref: "./_shared.yaml#/components/parameters/XUserId"
        - $ref: "./_shared.yaml#/components/parameters/XTenantId"
        - $ref: "./_shared.yaml#/components/parameters/XUserRole"
        - $ref: "./_shared.yaml#/components/parameters/XInternalService"
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        "201": { description: Recorded., content: { application/json: { schema: { $ref: "./_shared.yaml#/components/schemas/SuccessEnvelope" } } } }
        "403": { $ref: "./_shared.yaml#/components/responses/Forbidden" }
  /internal/applications/{id}:
    patch:
      tags: [applications]
      summary: Update application stage/status (ADMIN/RECRUITER/HIRING_MANAGER)
      parameters:
        - $ref: "./_shared.yaml#/components/parameters/XUserId"
        - $ref: "./_shared.yaml#/components/parameters/XTenantId"
        - $ref: "./_shared.yaml#/components/parameters/XUserRole"
        - $ref: "./_shared.yaml#/components/parameters/XInternalService"
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        "200": { description: Updated., content: { application/json: { schema: { $ref: "./_shared.yaml#/components/schemas/SuccessEnvelope" } } } }
        "403": { $ref: "./_shared.yaml#/components/responses/Forbidden" }
  /internal/applications/{id}/hire:
    post:
      tags: [applications]
      summary: Hire (idempotent; publishes application.hired)
      parameters:
        - $ref: "./_shared.yaml#/components/parameters/XUserId"
        - $ref: "./_shared.yaml#/components/parameters/XTenantId"
        - $ref: "./_shared.yaml#/components/parameters/XUserRole"
        - $ref: "./_shared.yaml#/components/parameters/XInternalService"
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: false
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        "200": { description: Hired (no re-stamp if already hired)., content: { application/json: { schema: { $ref: "./_shared.yaml#/components/schemas/SuccessEnvelope" } } } }
        "403": { $ref: "./_shared.yaml#/components/responses/Forbidden" }
  /internal/applications/{id}/reject:
    post:
      tags: [applications]
      summary: Reject (publishes application.rejected)
      parameters:
        - $ref: "./_shared.yaml#/components/parameters/XUserId"
        - $ref: "./_shared.yaml#/components/parameters/XTenantId"
        - $ref: "./_shared.yaml#/components/parameters/XUserRole"
        - $ref: "./_shared.yaml#/components/parameters/XInternalService"
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: false
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        "200": { description: Rejected., content: { application/json: { schema: { $ref: "./_shared.yaml#/components/schemas/SuccessEnvelope" } } } }
        "403": { $ref: "./_shared.yaml#/components/responses/Forbidden" }
  # ─── Offers (routes/offers.ts) ────────────────────────────────────────────
...